requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
